test/requirements.txt
streamlit==1.25.0
tensorflow==2.19.0
keras==3.10.0
pillow
numpy